Refactoring: move CLI apps to gemini.cmd package
