+++ report
@@ -8,7 +8,7 @@ jar="target/gemini-${version}.jar"
 build_command="./sbt package"
 current_dir="$(dirname "$0")"
 
-app_class="tech.sourced.gemini.ReportApp"
+app_class="tech.sourced.gemini.cmd.ReportApp"
 
 hash java >/dev/null 2>&1 || { echo "Please install Java" >&2; exit 1; }
